Cohort Characterization with Strata fails with temp table already exists

+++ src/main/java/org/ohdsi/webapi/service/AbstractDaoService.java
@@ -58,6 +58,9 @@ public abstract class AbstractDaoService {
   @Value("${cdm.version}")
   private String cdmVersion;
 
+  @Value("${jdbc.suppressInvalidApiException}")
+  private boolean suppressApiException;
+
   @Autowired
   private JdbcTemplate jdbcTemplate;
 
@@ -155,7 +158,9 @@ public abstract class AbstractDaoService {
     } else {
       dataSource = new DriverManagerDataSource(dataSourceData.getConnectionString());
     }
-    return new CancelableJdbcTemplate(dataSource);
+    CancelableJdbcTemplate jdbcTemplate = new CancelableJdbcTemplate(dataSource);
+    jdbcTemplate.setSuppressApiException(suppressApiException);
+    return jdbcTemplate;
   }
 
   private void loginToKerberos(DataSourceUnsecuredDTO dataSourceData) {
@@ -169,7 +174,7 @@ public abstract class AbstractDaoService {
     }
     try {
       FileUtils.forceDelete(temporaryDir);
-      if (StringUtils.isNotBlank(krbConfig.getComponents().getKeytabPath().toString())){
+      if (Objects.nonNull(krbConfig.getComponents()) && StringUtils.isNotBlank(krbConfig.getComponents().getKeytabPath().toString())){
         FileUtils.forceDelete(krbConfig.getComponents().getKeytabPath().toFile());
       }
     } catch (IOException e) {
